searcher: renamed component
Renames component `search` to `searcher` for better consistency with
other components (`downloader`, `validator`, `verifier`, `validator`).

Fixes some flake8 warnings.
